+++ source/prepend.spec.js
@@ -1,16 +1,18 @@
 import { prepend } from './prepend'
 
 test('happy', () => {
-  expect(prepend('f', 'oo')).toEqual('foo')
-})
-
-test('prepend', () => {
   expect(prepend('yes', [ 'foo', 'bar', 'baz' ])).toEqual([
     'yes',
     'foo',
     'bar',
     'baz',
   ])
+})
 
+test('with empty list', () => {
   expect(prepend('foo')([])).toEqual([ 'foo' ])
+})
+
+test('with string instead of array', () => {
+  expect(prepend('foo')('bar')).toEqual([ 'foo', 'b', 'a', 'r' ])
 })