+++ scripts/populate-readme-data/assets/CHANGELOG.md
@@ -8,6 +8,8 @@ WIP 6.2.0
 
 - Add `R.splitWhen`
 
+- `R.append`/`R.prepend` now work only with arrays just like Ramda. Previous behaviour was for them to work with both arrays and strings.
+  
 6.1.0
 
 - Fix `R.and` wrong definition, because the function doesn't convert the result to boolean. This introduce another difference with `@types/ramda`.